A draft rpoposal to refactor `BatchSignatureVerifier`.
