+++ ethereum/spec/src/main/java/tech/pegasys/teku/spec/logic/common/util/AttestationUtil.java
@@ -215,7 +215,8 @@ public abstract class AttestationUtil {
    */
   public AttestationProcessingResult isValidIndexedAttestation(
       final Fork fork, final BeaconState state, final IndexedAttestation indexedAttestation) {
-    return isValidIndexedAttestation(fork, state, indexedAttestation, BLSSignatureVerifier.SIMPLE);
+    return isValidIndexedAttestation(
+        fork, state, indexedAttestation, specConfig.getBLSSignatureVerifier());
   }
 
   public AttestationProcessingResult isValidIndexedAttestation(
