+++ infrastructure/bls/src/main/java/tech/pegasys/teku/bls/BatchSignatureVerifier.java
@@ -11,7 +11,7 @@
  * specific language governing permissions and limitations under the License.
  */
 
-package tech.pegasys.teku.spec.logic.common.statetransition.blockvalidator;
+package tech.pegasys.teku.bls;
 
 import static com.google.common.base.Preconditions.checkArgument;
 
@@ -19,11 +19,6 @@ import com.google.common.annotations.VisibleForTesting;
 import java.util.ArrayList;
 import java.util.List;
 import org.apache.tuweni.bytes.Bytes;
-import tech.pegasys.teku.bls.BLS;
-import tech.pegasys.teku.bls.BLSPublicKey;
-import tech.pegasys.teku.bls.BLSSignature;
-import tech.pegasys.teku.bls.BLSSignatureVerifier;
-import tech.pegasys.teku.bls.BatchSemiAggregate;
 
 /**
  * Implementation which doesn't perform any actual validations on {@link #verify(List, Bytes,
@@ -35,7 +30,7 @@ import tech.pegasys.teku.bls.BatchSemiAggregate;
  *
  * <p>This is thread-safe class.
  */
-public class BatchSignatureVerifier implements BLSSignatureVerifier {
+public class BatchSignatureVerifier implements BatchBLSSignatureVerifier {
 
   private static class Job {
     final int idx;
@@ -90,6 +85,7 @@ public class BatchSignatureVerifier implements BLSSignatureVerifier {
    * <p>After this method completes the instance should be disposed and any subsequent calls to this
    * instance methods would fail with exception
    */
+  @Override
   public synchronized boolean batchVerify() {
     if (complete) {
       throw new IllegalStateException("Reuse of disposable instance");
