+++ infrastructure/bls/src/test/java/tech/pegasys/teku/bls/BatchSignatureVerifierTest.java
@@ -11,7 +11,7 @@
  * specific language governing permissions and limitations under the License.
  */
 
-package tech.pegasys.teku.spec.logic.common.statetransition.blockvalidator;
+package tech.pegasys.teku.bls;
 
 import static java.nio.charset.StandardCharsets.UTF_8;
 import static org.assertj.core.api.Assertions.assertThat;
@@ -26,9 +26,6 @@ import java.util.concurrent.Future;
 import java.util.stream.IntStream;
 import org.apache.tuweni.bytes.Bytes;
 import org.junit.jupiter.api.Test;
-import tech.pegasys.teku.bls.BLSPublicKey;
-import tech.pegasys.teku.bls.BLSSignature;
-import tech.pegasys.teku.bls.BLSTestUtil;
 
 public class BatchSignatureVerifierTest {
